Adding final files
Made many changes while trying to make the site live that were not reflected here on this repository. Temporarily had a github pages free hosting repository but that did not work with form submission. Hostinger free hosting didn't work either. Now that the site is finished and being hosted on (paid for) Hostinger, I'm updating the files here.